Updated the expand_summary_infChain function
Chain is expanded ID numbers min(infChain$i):max(chain$i) so that it can be used for a subset of the history matrix.

Chain is expanded for the unique sampno's to allow for uneven sampno's from thinning.

Chain can be expanded for a user given j, again to allow for subsetting. i.e. pass a 'global' j_vec for 'local' ids, in case some individuals are not alive/sampled for some j's.