Part way through adding stratification to kinetics parameters
I have functions in the test_model script which create a parameter table with an entry for each unique demographic grouping. Need to integrate in create_posterior_func. The idea is that we have a table of parameters, with one entry for each parameter name/group combination. It's easier to enumerate a whole table and have many dummy entries (i.e., same for all groups).